requirements.txt
google-genai
Pillow
numpy
opencv-python
flask
flask-socketio
flask-cors
azure-cognitiveservices-speech
